+++ lib/shared/services/permission_enum.dart
@@ -1,9 +1,9 @@
 /// lib/shared/permission_enum.dart
 /// for consistent implementation of the app
-/// we will define the consts or enum for Permission that might be reusable here
+/// we will define the enum for Permission that might be reusable here
 ///
 
-enum PermissionConst {
+enum PermissionEnum {
   photos,
   locationWhenInUse,
   calendarFullAccess,